added to commentProductFragment and viewModel, click comment features

+++ app/src/main/java/com/skapps/shoppingapp/data/responce/CommentResponce.kt
@@ -5,6 +5,6 @@ import com.google.gson.annotations.SerializedName
 data class CommentResponce(
     @SerializedName("header"  ) var header  : String? = null,
     @SerializedName("content" ) var content : String? = null,
-    @SerializedName("rating"  ) var rating  : Int?    = null,
+    @SerializedName("rating"  ) var rating  : Double?    = null,
     @SerializedName("status"  ) var status  : String? = null
 )
